+++ src/lib.rs
@@ -25,5 +25,5 @@ pub mod weights;
 
 #[cfg(feature = "std")]
 pub use decoder::Decoder;
-pub use decoder::output::RgbImage;
+pub use decoder::output::{Picture, RgbImage, YuvImage};
 pub use error::Error;